Split implementation out from PlatformWebView on WinForms.
