+++ src/Auth0.OidcClient.WinForms/Auth0.OidcClient.WinForms.csproj
@@ -52,6 +52,7 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="PlatformWebView.cs" />
+    <Compile Include="WebBrowserBrowser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
